+++ src/StaticHelper/ApiHelper.cs
@@ -251,98 +251,5 @@ namespace StaticHelper
                 .ToString(ISOCode.Language.L___);
             return result;
         }
-
-        /// <summary>
-        /// Возвращает имя канала (IO-Link, DI, DO) из строки для IO-Link
-        /// модуля.
-        /// </summary>
-        /// <param name="source">Строка для поиска</param>
-        /// <returns></returns>
-        public static string GetChannelNameForIOLinkModuleFromString(
-            string source)
-        {
-            const string IOLink = "IO-Link";
-            const string DI = "DI";
-            const string DO = "DO";
-
-            if (source.Contains(DI) &&
-                !source.Contains(IOLink) &&
-                !source.Contains(DO))
-            {
-                return DI;
-            }
-
-            if (source.Contains(DO) &&
-                !source.Contains(IOLink) &&
-                !source.Contains(DI))
-            {
-                return DO;
-            }
-
-            return "";
-        }
-
-        /// <summary>
-        /// MatchEvaluator для regular expression,
-        /// замена русских букв на английские
-        /// </summary>
-        public static string RussianToEnglish(Match m)
-        {
-            switch (m.ToString()[0])
-            {
-                case 'А':
-                    return "A";
-                case 'В':
-                    return "B";
-                case 'С':
-                    return "C";
-                case 'Е':
-                    return "E";
-                case 'К':
-                    return "K";
-                case 'М':
-                    return "M";
-                case 'Н':
-                    return "H";
-                case 'Х':
-                    return "X";
-                case 'Р':
-                    return "P";
-                case 'О':
-                    return "O";
-                case 'Т':
-                    return "T";
-            }
-
-            return m.ToString();
-        }
-
-        public static string GetClampNumberAsString(
-            Function deviceClampFunction)
-        {
-            string clampNumberAsString = deviceClampFunction.Properties
-                .FUNC_ADDITIONALIDENTIFYINGNAMEPART.ToString();
-
-            bool haveValveTerminal = deviceClampFunction.Name
-                .Contains(Device.DeviceManager.ValveTerminalName);
-            if (haveValveTerminal)
-            {
-                Function IOModuleFunction = 
-                    GetIOModuleFunction(deviceClampFunction);
-                string bindedDevice = deviceClampFunction.Name;
-                Function IOModuleClampFunction = 
-                    GetClampFunction(IOModuleFunction, bindedDevice);
-                clampNumberAsString = IOModuleClampFunction.Properties
-                .FUNC_ADDITIONALIDENTIFYINGNAMEPART.ToString();
-            }
-
-            return clampNumberAsString ?? string.Empty;
-        }
-
-        public static string ReplaceRusBigLettersByEngBig(string replacingStr)
-        {
-            return Regex.Replace(replacingStr,
-                CommonConst.RusAsEngPattern, CommonConst.RusAsEngEvaluator);
-        }
     }
 }
